examples/mobc: apply rename migration(Applications -> applications)
