--- examples/mobc/src/src_user/Applications/app_headers.h
@@ -1,31 +1,0 @@
-/**
- * @file
- * @brief Appのヘッダをまとめたヘッダ
- */
-#ifndef APP_HEADERS_H_
-#define APP_HEADERS_H_
-
-// Core
-#include <src_core/Applications/nop.h>
-#include <src_core/Applications/gs_command_dispatcher.h>
-#include <src_core/Applications/realtime_command_dispatcher.h>
-#include <src_core/Applications/timeline_command_dispatcher.h>
-#include <src_core/Applications/event_utility.h>
-#include <src_core/Applications/memory_dump.h>
-#include <src_core/Applications/telemetry_manager.h>
-#include <src_core/Applications/utility_command.h>
-#include <src_core/Applications/utility_counter.h>
-#include <src_core/Applications/divided_cmd_utility.h>
-
-// Core TestApp
-#include <src_core/Applications/TestApp/test_ccp_util.h>
-
-// DI
-#include "DriverInstances/di_aobc.h"
-#include "DriverInstances/di_uart_test.h"
-#include "DriverInstances/di_gs.h"
-
-// UserDefined
-#include "UserDefined/debug_apps.h"
-
-#endif
