+++ examples/mobc/src/src_user/Settings/Modes/Transitions/sequence_items.h
@@ -2,7 +2,7 @@
 #define SEQUENCE_ITEMS_H_
 
 #include <src_core/TlmCmd/block_command_table.h>
-#include <src_core/Applications/timeline_command_dispatcher_id_define.h>
+#include <src_core/applications/timeline_command_dispatcher_id_define.h>
 
 // !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 // SIを使うことはもう非推奨！！！
