+++ examples/mobc/src/src_user/Settings/Modes/Transitions/sl_initial.c
@@ -6,7 +6,7 @@
 
 #include "../../../Drivers/Com/gs.h"
 
-#include <src_core/Applications/timeline_command_dispatcher_id_define.h>
+#include <src_core/applications/timeline_command_dispatcher_id_define.h>
 #include <src_core/TlmCmd/block_command_loader.h>
 #include <src_core/System/TimeManager/obc_time.h>
