+++ examples/mobc/src/src_user/Settings/applications/telemetry_manager_define.h
