+++ examples/mobc/src/src_user/TlmCmd/NormalBlockCommandDefinition/nbc_start_hk_tlm.c
@@ -5,7 +5,7 @@
 #include "../command_definitions.h"
 #include "../telemetry_definitions.h"
 
-#include <src_core/Applications/timeline_command_dispatcher_id_define.h>
+#include <src_core/applications/timeline_command_dispatcher_id_define.h>
 #include <src_core/TlmCmd/block_command_loader.h>
