+++ examples/mobc/src/src_user/applications/app_headers.h
@@ -1,0 +1,31 @@
+/**
+ * @file
+ * @brief Appのヘッダをまとめたヘッダ
+ */
+#ifndef APP_HEADERS_H_
+#define APP_HEADERS_H_
+
+// Core
+#include <src_core/applications/nop.h>
+#include <src_core/applications/gs_command_dispatcher.h>
+#include <src_core/applications/realtime_command_dispatcher.h>
+#include <src_core/applications/timeline_command_dispatcher.h>
+#include <src_core/applications/event_utility.h>
+#include <src_core/applications/memory_dump.h>
+#include <src_core/applications/telemetry_manager.h>
+#include <src_core/applications/utility_command.h>
+#include <src_core/applications/utility_counter.h>
+#include <src_core/applications/divided_cmd_utility.h>
+
+// Core TestApp
+#include <src_core/applications/test_app/test_ccp_util.h>
+
+// DI
+#include "driver_instances/di_aobc.h"
+#include "driver_instances/di_uart_test.h"
+#include "driver_instances/di_gs.h"
+
+// UserDefined
+#include "user_defined/debug_apps.h"
+
+#endif
