+++ examples/mobc/src/src_user/applications/app_registry.c
