+++ examples/mobc/src/src_user/applications/driver_instances/di_gs.c
